add: saving last read location

+++ app/src/main/java/com/example/reader_v2/data/entity/BookEntity.kt
@@ -13,7 +13,7 @@ data class BookEntity(
     val contributors: List<Contributor>?,
     val description: String?,
     val coverImagePath: String?,
-    val progression: Float,
+    val progression: Double,
     val lastReadLocation: String?,
     val addedDate: Long,
     val lastReadDate: Long
