+++ app/src/main/java/com/example/reader_v2/data/model/Book.kt
@@ -11,7 +11,7 @@ data class Book(
 	val contributors: List<Contributor>?,
 	val description: String?,
 	val coverImagePath: String?,
-	val progression: Float,
+	val progression: Double,
 	val lastReadLocation: String?,
 	val addedDate: Long,
 	val lastReadDate: Long
